JCF: manually suppress misleading complaint from clang-tidy about header files not being found
Frequently, when running clang-tidy on a file, I'll see an error like the following:

/home/jcfree/artdaq-demo_v3_08_00/products/artdaq_core/v3_05_14/include/artdaq-core/Data/Fragment.hh:4:10: error: 'algorithm' file not found [clang-diagnostic-error]
         ^

Apparently this can't be turned off via a simple clang-tidy option,
and it's (A) false, and (B) pointless even if it weren't false since
gcc will most assuredly get upset if it can't find a header. So I'm deleting this type of file not found error from the clang-tidy output via sed.