JCF: locally modify the compile_commands.json file so non-user headers don't create warnings
clang-tidy by default appears to go all the way up the #include tree when you have it parse a
source file, with the annoying result that it prints warnings about code in headers over which
the developer has no control. I've added logic to duneclang-tidy.sh so that it makes a copy of
the compile_commands.json file whose directory it's been given, and modifies that file so that
out warnings. Right now, the logic is that if an #include directory looks like this:

/../../../../<package version, e.g., v3_2_1>/../../include

...then it gets ignored. This means, e.g., ups package include directories are ignored. This
algorithm will likely need to be changed in the future, but it's a start.