JCF: remove requirement that increment or decrement be PREincrement or PREdecrement
As mentioned in the original Google Style Guide, preincrementing
rather than postincrementing tends to be more efficient as a copy
isn't made of the object being incremented. However, unless the value
from before the increment is used, compilers tend not to make the
copy, i.e., unless you do something like this:

int var1 = var2++;

And since rule specific to the DUNE Style Guide (if not the Google
one) is that increments/decrements be the only thing done on a line,
we don't really run this risk.