JCF: remove build/explicit_make_pair from dunecpplint
There was a bug when g++ 4.6 was in C++11 mode when make_pair's
template arguments were specified explicity. This is no longer the
case with newer versions of gcc, and furthermore, the Style Guide
actually encourages users to make template parameters in function
explicit even though they can be deduced by the compiler.